+++ src/components/Auth/Auth.scss
@@ -2,7 +2,7 @@
   background-color: white;
 }
 
-.login {
+.form {
   .submit-btn-wrapper {
     text-align: right;
     width: 256px;
